// src/lib/api/models/SmsHistoryPublic.ts
export interface SmsHistoryPublic {
    /**
     * 
     * @type {string}
     * @memberof SmsHistoryPublic
     */
    recipient: string;
    /**
     * 
     * @type {string}
     * @memberof SmsHistoryPublic
     */
    message: string;
    /**
     * 
     * @type {string}
     * @memberof SmsHistoryPublic
     */
    status?: string;
    /**
     * 
     * @type {number}
     * @memberof SmsHistoryPublic
     */
    smsCount?: number;
    /**
     * 
     * @type {number}
     * @memberof SmsHistoryPublic
     */
    cost: number;
    /**
     * 
     * @type {string}
     * @memberof SmsHistoryPublic
     */
    templateId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SmsHistoryPublic
     */
    errorMessage?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SmsHistoryPublic
     */
    deliveryStatus?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SmsHistoryPublic
     */
    externalId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SmsHistoryPublic
     */
    id: string;
    /**
     * 
     * @type {Date}
     * @memberof SmsHistoryPublic
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof SmsHistoryPublic
     */
    updatedAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof SmsHistoryPublic
     */
    sentAt: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof SmsHistoryPublic
     */
    deliveredAt: Date | null;
    /**
     * 
     * @type {string}
     * @memberof SmsHistoryPublic
     */
    userId: string;
}